+++ modules/cirrus/outputs.tf
@@ -92,6 +92,20 @@ output "cirrus_workflow_state_machine_role_arns" {
   }
 }
 
+output "cirrus_task_role_arns" {
+  description = <<-DESCRIPTION
+  Map of cirrus task names to their IAM role ARN.
+  DESCRIPTION
+
+  value = {
+    for task_name, task_output in module.task :
+    task_name => {
+      "lambda" = task_output.lambda.role_arn
+      "batch"  = task_output.batch.role_arn
+    }
+  }
+}
+
 output "cirrus_task_batch_compute_template_variables" {
   description = <<-DESCRIPTION
   Map of task batch compute names to their resolved template variables (static and SSM combined).