+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Added to TF outputs mapping of Cirrus lambda/batch tasks to IAM role arns
+
 ### Changed
 
 ### Fixed